Extract a static library for the common code between the 2 parts of the day 4.
