+++ 2015/Day4/CMakeLists.txt
@@ -21,15 +21,17 @@ conan_cmake_run(REQUIRES OpenSSL/1.1.1a@conan/stable
                 OPTIONS OpenSSL:shared=False
 )
 
-message ("Conan libs : ${CONAN_LIBS}")
+add_library(Md5Wrapper STATIC src/md5_wrapper.cpp)
+set_target_properties(Md5Wrapper PROPERTIES CXX_STANDARD 17)
+target_link_libraries(Md5Wrapper PUBLIC CONAN_PKG::OpenSSL)
 
 add_executable(TheIdealStockingStuffer_Part1 src/part_one.cpp)
 set_target_properties(TheIdealStockingStuffer_Part1 PROPERTIES CXX_STANDARD 17)
-target_link_libraries(TheIdealStockingStuffer_Part1 PUBLIC CONAN_PKG::OpenSSL)
+target_link_libraries(TheIdealStockingStuffer_Part1 PUBLIC Md5Wrapper)
 
 add_executable(TheIdealStockingStuffer_Part2 src/part_two.cpp)
 set_target_properties(TheIdealStockingStuffer_Part2 PROPERTIES CXX_STANDARD 17)
-target_link_libraries(TheIdealStockingStuffer_Part2 PUBLIC CONAN_PKG::OpenSSL)
+target_link_libraries(TheIdealStockingStuffer_Part2 PUBLIC Md5Wrapper)
 
 add_test(NAME TheIdealStockingStuffer_part_one_example_1
     COMMAND $<TARGET_FILE:TheIdealStockingStuffer_Part1> "abcdef" 609043
