+++ 2015/Day4/src/md5_wrapper.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string>
+
+std::string getMD5(std::string key);
